pyproject.toml
[project]
name = "restapi-application"
version = "0.1.0"
description = ""
authors = [
    {name = "Maksym Kryvenko",email = "maks1997ym@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic[email] (>=2.13.3,<3.0.0)",
    "fastapi (>=0.136.0,<0.137.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)",
    "bcrypt",
    "python-jose[cryptography]",
    "fastapi-mail",
    "cloudinary",
    "redis",
    "libgravatar",
    "python-multipart",
    "slowapi ; python_version >= '3.14' and python_version < '4.0'",
    "pytest (>=9.0.3,<10.0.0)",
    "sphinx (>=9.1.0,<10.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "black (>=26.3.1,<27.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
filterwarnings = "ignore::DeprecationWarning"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
